requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyJWT==2.9.0
pydantic==2.8.2
email-validator==2.2.0
python-multipart==0.0.9
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2
pypdf==4.3.1
python-docx==1.1.2
prometheus-client==0.20.0
openai>=1.50.0
